python/iflearner_flow_client/requirements.txt
click==8.0
PyYAML==6.0
setuptools
